+++ osu.Game.Rulesets.Karaoke/KaraokeControlInputManager.cs
@@ -1,0 +1,61 @@
+﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+using System.ComponentModel;
+
+namespace osu.Game.Rulesets.Karaoke
+{
+    public enum KaraokeAction
+    {
+        [Description("First Lyric")]
+        FirstLyric,
+
+        [Description("Previous BaseLyric")]
+        PreviousLyric,
+
+        [Description("Next BaseLyric")]
+        NextLyric,
+
+        [Description("Play and pause")]
+        PlayAndPause,
+
+        [Description("Open/Close adjustment")]
+        OpenPanel,
+
+        [Description("Increase Speed")]
+        IncreaseTempo,
+
+        [Description("Decrease Speed")]
+        DecreaseTempo,
+
+        [Description("Reset Speed")]
+        ResetTempo,
+
+        [Description("Increase pitch")]
+        IncreasePitch,
+
+        [Description("Decrease pitch")]
+        DecreasePitch,
+
+        [Description("Reset pitch")]
+        ResetPitch,
+
+        [Description("Increase vocal pitch")]
+        IncreaseVocalPitch,
+
+        [Description("Decrease vocal pitch")]
+        DecreaseVocalPitch,
+
+        [Description("Reset vocal pitch")]
+        ResetVocalPitch,
+
+        [Description("Increase saiten pitch")]
+        IncreaseSaitenPitch,
+
+        [Description("Decrease saiten pitch")]
+        DecreaseSaitenPitch,
+
+        [Description("Reset saiten pitch")]
+        ResetSaitenPitch,
+    }
+}
