+++ osu.Game.Rulesets.Karaoke/KaraokeInputManager.cs
@@ -154,58 +154,4 @@ namespace osu.Game.Rulesets.Karaoke
     {
         public float Scale { get; set; }
     }
-
-    public enum KaraokeAction
-    {
-        [Description("First Lyric")]
-        FirstLyric,
-
-        [Description("Previous BaseLyric")]
-        PreviousLyric,
-
-        [Description("Next BaseLyric")]
-        NextLyric,
-
-        [Description("Play and pause")]
-        PlayAndPause,
-
-        [Description("Open/Close adjustment")]
-        OpenPanel,
-
-        [Description("Increase Speed")]
-        IncreaseTempo,
-
-        [Description("Decrease Speed")]
-        DecreaseTempo,
-
-        [Description("Reset Speed")]
-        ResetTempo,
-
-        [Description("Increase pitch")]
-        IncreasePitch,
-
-        [Description("Decrease pitch")]
-        DecreasePitch,
-
-        [Description("Reset pitch")]
-        ResetPitch,
-
-        [Description("Increase vocal pitch")]
-        IncreaseVocalPitch,
-
-        [Description("Decrease vocal pitch")]
-        DecreaseVocalPitch,
-
-        [Description("Reset vocal pitch")]
-        ResetVocalPitch,
-
-        [Description("Increase saiten pitch")]
-        IncreaseSaitenPitch,
-
-        [Description("Decrease saiten pitch")]
-        DecreaseSaitenPitch,
-
-        [Description("Reset saiten pitch")]
-        ResetSaitenPitch,
-    }
 }
